Implemented it so that a random song will play when the app is opened. (ChooseRandomSongActivity)
Implemented music seek bar? (SetMusicSliderActivity)Double check this.

Implemented basics for sounds seek bar. (SetSoundSliderActivity)
We need to be farther into this project where we are implementing sounds that go with the snake character before we can finish this

Note to self - figure out emulator problem so I can test this stuff myself